Add Google OAuth

+++ controllers/api/v1/posts_api.js
@@ -19,11 +19,18 @@ module.exports.destroy = async (req, res) => {
 	try {
 		let post = await Post.findById(req.params.id);
 
-		post.remove();
-		await Comment.deleteMany({ post: req.params.id });
-		return res.json(200, {
-			message : 'Post and associated comments deleted'
-		});
+		if (post.user == req.user.id) {
+			post.remove();
+			await Comment.deleteMany({ post: req.params.id });
+			return res.json(200, {
+				message : 'Post and associated comments deleted'
+			});
+		}
+		else {
+			return res.json(401, {
+				message : 'Unauthorized - Cannot delete this post'
+			});
+		}
 	} catch (err) {
 		return res.json(500, {
 			message : 'Internal Server Error'
